refactor(fxa-shared): cleanup sentry and add exc route
Because:

* Sentry relocation didn't clean up reporting functionality from the
  service using it.
* We had no way to verify exceptions are thrown and captured by Sentry.

This commit:

* Refactors the sentry reporting functionality so it can be used
  directly by other code.
* Adds an __exception__ route with basic trigger protection to avoid
  excessive error reporting.

Closes #6454
