+++ packages/fxa-shared/nestjs/sentry/sentry.interceptor.ts
@@ -7,45 +7,10 @@ import {
   Injectable,
   NestInterceptor,
 } from '@nestjs/common';
-import { GqlContextType, GqlExecutionContext } from '@nestjs/graphql';
-import * as Sentry from '@sentry/node';
-import { Request } from 'express';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-function processException(context: ExecutionContext, exception: Error) {
-  // First determine what type of a request this is
-  let requestType: 'http' | 'graphql' | undefined;
-  let request: Request;
-  let gqlExec: GqlExecutionContext | undefined;
-  if (context.getType() === 'http') {
-    requestType = 'http';
-    request = context.switchToHttp().getRequest();
-  } else if (context.getType<GqlContextType>() === 'graphql') {
-    requestType = 'graphql';
-    gqlExec = GqlExecutionContext.create(context);
-    request = gqlExec.getContext().req;
-  }
-
-  Sentry.withScope((scope: Sentry.Scope) => {
-    scope.addEventProcessor((event: Sentry.Event) => {
-      if (requestType) {
-        const sentryEvent = Sentry.Handlers.parseRequest(event, request);
-        sentryEvent.level = Sentry.Severity.Error;
-        return sentryEvent;
-      }
-      return null;
-    });
-    if (gqlExec) {
-      const info = gqlExec.getInfo();
-      scope.setContext('graphql', {
-        fieldName: info.fieldName,
-        path: info.path,
-      });
-    }
-    Sentry.captureException(exception);
-  });
-}
+import { processException } from './reporting';
 
 @Injectable()
 export class SentryInterceptor implements NestInterceptor {
