+++ packages/fxa-shared/nestjs/sentry/sentry.service.ts
@@ -4,22 +4,14 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { ExtraErrorData } from '@sentry/integrations';
 import * as Sentry from '@sentry/node';
-import { SQS } from 'aws-sdk';
 
+import { filterSentryEvent } from './reporting';
 import { SENTRY_CONFIG } from './sentry.constants';
 import { SentryConfigParams } from './sentry.module';
 
-// Matches uid, session, oauth and other common tokens which we would
-// prefer not to include in Sentry reports.
-const TOKENREGEX = /[a-fA-F0-9]{32,}/gi;
-// RFC 5322 generalized email regex, ~ 99.99% accurate.
-const EMAILREGEX = /(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))/gi;
-const FILTERED = '[Filtered]';
-const URIENCODEDFILTERED = encodeURIComponent(FILTERED);
-
 @Injectable()
 export class SentryService {
-  constructor(@Inject(SENTRY_CONFIG) private sentryConfig: SentryConfigParams) {
+  constructor(@Inject(SENTRY_CONFIG) sentryConfig: SentryConfigParams) {
     // Setup Sentry
     Sentry.init({
       ...sentryConfig,
@@ -29,94 +21,4 @@ export class SentryService {
       },
     });
   }
-}
-
-/**
- * Filters all of an objects string properties to remove tokens.
- *
- * @param obj Object to filter values on
- */
-function filterObject<T>(obj: T): T {
-  if (typeof obj === 'object' && obj) {
-    for (const [key, value] of Object.entries(obj)) {
-      if (typeof value === 'string') {
-        // Typescript can't quite infer that this is the value that was
-        // at that index, so a cast is needed.
-        (obj as any)[key] = value
-          .replace(TOKENREGEX, FILTERED)
-          .replace(EMAILREGEX, FILTERED);
-      }
-    }
-  }
-  return obj;
-}
-
-/**
- * Filter a sentry event for PII in addition to the default filters.
- *
- * Current replacements:
- *   - A 32-char hex string that typically is a FxA user-id.
- *
- * Data Removed:
- *   - Request body.
- *
- * @param event
- */
-function filterSentryEvent(event: Sentry.Event, hint: unknown) {
-  if (event.message) {
-    event.message = event.message.replace(TOKENREGEX, FILTERED);
-  }
-  if (event.breadcrumbs) {
-    for (const bc of event.breadcrumbs) {
-      if (bc.message) {
-        bc.message = bc.message.replace(TOKENREGEX, FILTERED);
-      }
-      if (bc.data) {
-        bc.data = filterObject(bc.data);
-      }
-    }
-  }
-  if (event.request) {
-    if (event.request.url) {
-      event.request.url = event.request.url.replace(TOKENREGEX, FILTERED);
-    }
-    if (event.request.query_string) {
-      event.request.query_string = event.request.query_string.replace(
-        TOKENREGEX,
-        URIENCODEDFILTERED
-      );
-    }
-    if (event.request.headers) {
-      (event as any).request.headers = filterObject(event.request.headers);
-    }
-    if (event.request.data) {
-      // Remove request data entirely
-      delete event.request.data;
-    }
-  }
-  if (event.tags && event.tags.url) {
-    event.tags.url = event.tags.url.replace(TOKENREGEX, FILTERED);
-  }
-  return event;
-}
-
-/**
- * Capture a SQS Error to Sentry with additional context.
- *
- * @param err Error object to capture.
- * @param message SQS Message to include with error.
- */
-export function captureSqsError(err: Error, message?: SQS.Message): void {
-  Sentry.withScope((scope) => {
-    if (message?.Body) {
-      if (typeof message.Body === 'string') {
-        message.Body = message.Body.replace(TOKENREGEX, FILTERED).replace(
-          EMAILREGEX,
-          FILTERED
-        );
-      }
-      scope.setContext('SQS Message', message);
-    }
-    Sentry.captureException(err);
-  });
 }
